Add functionality to upload release asset
This commit introduces a new step to the Github workflow - the uploading of the release asset. It does this by creating a zip file from the build directory and uploading it to the appropriate release on Github. This allows for the build to be easily accessed and downloaded as part of each release.