Improve GitHub Actions workflow and project versioning
This commit introduces improvements in the automated process for the CI/CD pipeline using GitHub actions and modifies the projects versioning. It includes the addition of PowerShell scripts for creating a GitHub release and uploading the release asset. The Continuous Integration configuration has been updated to call these scripts instead. Also, a process has been added to extract the version number from the project file. This significantly enhances the deployment process and ensures a clean, efficient mechanism for software releases.