Update version and refactor component properties handling
Relocated the Main method for improved readability and streamlined the process of retrieving component properties by refining the handling of null or whitespace values. Logging calls in the search for component details were purged. Also, updated the TryGetValue method to set the value parameter to null, rather than an empty string, when there's no return. The version in DetectiveSpecs.csproj was also incremented.