Move Component.cs and refactor code in Program.cs
Component.cs was moved from the Models directory to the DetectiveSpecs root directory. Additionally, the code in Program.cs was refactored to encapsulate functionality to read component properties into a separate method and the Main method was updated to change the files that are generated. This change makes the code more maintainable and easy to understand.