Refactor ComputerSpecSerializer and update assemblies namespace
ComputerSpecSerializer is refactored from a static class into an instance class to better manage the utilization of string builders and pad length. Also, namespaces of multiple assemblies are updated from `DetectiveSpecs.Models` to `DetectiveSpecs.Enums` to reflect the structural changes in the project. The version is also updated from 1.0.6 to 1.0.7.