Refactor component serialization methods in DetectiveSpecs
The Component class has been updated to include ComponentType. This change affected serialization methods in ComputerSpecSerializer.cs which got rid of an unneeded parameter. The code is now cleaner and easier to maintain. It has also led to modifications in Program.cs to adjust the way Component objects are created.